+++ R/saitem.R
@@ -3,9 +3,9 @@ NULL
 
 #' Read SAItem
 #'
-#' `.jsa_results()` extracts specific variables of the model of the SAItem while
-#' `.jsa_read()` extracts all the informations of a SAItem (see details).
-#'  `.jsa_jresults()` extracts the Java object of the results of a SAItem.
+#' `.jsai_results()` extracts specific variables of the model of the SAItem while
+#' `read_sai()` extracts all the informations of a SAItem (see details).
+#'  `.jsai_jresults()` extracts the Java object of the results of a SAItem.
 #'
 #' @param jsa Java SAItem object.
 #' @param items vector of characters containing the variables to extract.
@@ -13,7 +13,7 @@ NULL
 #' By default, extracts all the possible variables.
 #'
 #' @details A SAItem contains more information than just the results of a model.
-#' All those informations are extracted with the `.jsa_read()` function that
+#' All those informations are extracted with the `read_sai()` function that
 #' returns a list with 5 objects:
 #'
 #' - `ts`: the raw time series.
@@ -24,7 +24,7 @@ NULL
 #' estimation (fully identified model).
 #' - `results`: the result of the model.
 #' @export
-.jsa_read <- function(jsa) {
+read_sai <- function(jsa) {
     #  if (! .jcall(jsa, "Z", "isProcessed"))
     #    stop("You must run '.jws_compute()' on your workspace.")
 
@@ -33,7 +33,11 @@ NULL
     jestimation <- .jcall(jsa, "Ljdplus/sa/base/api/SaEstimation;", "getEstimation")
     jrslt <- .jnull()
     if (!is.jnull(jestimation)) {
-        jrslt <- .jcall(jestimation, "Ljdplus/toolkit/base/api/information/Explorable;", "getResults")
+        jrslt <- .jcall(
+            obj = jestimation,
+            returnSig = "Ljdplus/toolkit/base/api/information/Explorable;",
+            method = "getResults"
+        )
     }
     # ts
     jts <- .jcall(jdef, "Ljdplus/toolkit/base/api/timeseries/Ts;", "getTs")
@@ -88,9 +92,9 @@ NULL
     ))
 }
 
-#' @name .jsa_read
+#' @name read_sai
 #' @export
-.jsa_results <- function(jsa, items = NULL) {
+.jsai_results <- function(jsa, items = NULL) {
     jestimation <- .jcall(jsa, "Ljdplus/sa/base/api/SaEstimation;", "getEstimation")
     if (is.jnull(jestimation)) {
         return(NULL)
@@ -106,9 +110,9 @@ NULL
     return(r)
 }
 
-#' @name .jsa_read
+#' @name read_sai
 #' @export
-.jsa_jresults <- function(jsa) {
+.jsai_jresults <- function(jsa) {
     jestimation <- .jcall(jsa, "Ljdplus/sa/base/api/SaEstimation;", "getEstimation")
     if (is.jnull(jestimation)) {
         return(NULL)
@@ -121,7 +125,7 @@ NULL
 
 #' @name .jsap_name
 #' @export
-.jsa_name <- function(jsa) {
+.jsai_name <- function(jsa) {
     return(.jcall(jsa, "S", "getName"))
 }
 
@@ -129,17 +133,27 @@ NULL
 #'
 #' Extract specific metadata or time series metadata of a SAItem.
 #'
-#' @inheritParams .jsa_read
+#' @inheritParams read_sai
 #' @param key key of the metadata.
 #' @export
-.jsa_metadata <- function(jsa, key) {
-    val <- .jcall("jdplus/sa/base/workspace/Utility", "S", "getSingleMetaData", jsa, as.character(key))
+.jsai_metadata <- function(jsa, key) {
+    val <- .jcall(
+        obj = "jdplus/sa/base/workspace/Utility",
+        returnSig = "S",
+        method = "getSingleMetaData",
+        jsa, as.character(key)
+    )
     return(val)
 }
 
-#' @name .jsa_metadata
+#' @name .jsai_metadata
 #' @export
-.jsa_ts_metadata <- function(jsa, key) {
-    val <- .jcall("jdplus/sa/base/workspace/Utility", "S", "getSingleTsMetaData", jsa, as.character(key))
+.jsai_ts_metadata <- function(jsa, key) {
+    val <- .jcall(
+        obj = "jdplus/sa/base/workspace/Utility",
+        returnSig = "S",
+        method = "getSingleTsMetaData",
+        jsa, as.character(key)
+    )
     return(val)
 }